+++ botrunner/internal/game/reward.go
@@ -4,6 +4,6 @@ type Reward struct {
 	Id       int
 	Name     string
 	Type     string
-	Amount   float32
+	Amount   int64
 	Schedule string
 }
